feat: 지수 정보 엔티티 구현, DDL 작성, 프로젝트 구조 변경
--- 본문 ---
- 지수 정보 엔티티 구현
- 엔티티 공통 부분 구현
- layered-architecture 구조로 변경
- DDL(resources/db/schema.sql) 작성

--- 꼬리말 ---
close: #33
