+++ build.gradle
@@ -37,6 +37,7 @@ dependencies {
   implementation('org.springframework.boot:spring-boot-starter-validation')
   implementation('org.springdoc:springdoc-openapi-starter-webmvc-ui')
   implementation('org.springframework.boot:spring-boot-starter-data-jpa')
+  implementation 'org.springframework.boot:spring-boot-starter-webflux'
 
   implementation('me.paulschwarz:spring-dotenv:3.0.0')
