+++ src/main/java/com/sprint/project/findex/Findex.java
@@ -2,9 +2,7 @@ package com.sprint.project.findex;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@EnableJpaAuditing
 @SpringBootApplication
 public class Findex {
