+++ src/main/java/com/sprint/project/findex/dto/IndexInfoDto.java
@@ -1,0 +1,5 @@
+package com.sprint.project.findex.dto;
+
+public record IndexInfoDto() {
+
+}
