+++ src/main/java/com/sprint/project/findex/entity/SourceType.java
@@ -1,0 +1,6 @@
+package com.sprint.project.findex.entity;
+
+public enum SourceType {
+  USER,
+  OPEN_API
+}
