+++ src/main/java/com/sprint/project/findex/entity/base/BaseEntity.java
@@ -1,0 +1,46 @@
+package com.sprint.project.findex.entity.base;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+import java.io.Serial;
+import java.io.Serializable;
+import java.time.Instant;
+import java.util.Objects;
+import java.util.function.Consumer;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseEntity implements Serializable {
+
+  @Serial
+  private static final long serialVersionUID = 2L;
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @CreatedDate
+  @Column(updatable = false, nullable = false)
+  private Instant createdAt;
+
+  @LastModifiedDate
+  @Column
+  private Instant updatedAt;
+
+  protected <T> boolean updateIfChanged(T before, T after, Consumer<T> action) {
+    if (Objects.isNull(after) || Objects.equals(before, after)) {
+      return false;
+    }
+    action.accept(after);
+    return true;
+  }
+}
