--- src/main/java/com/sprint/project/findex/global/entity/SourceType.java
@@ -1,6 +1,0 @@
-package com.sprint.project.findex.global.entity;
-
-public enum SourceType {
-  USER,
-  OPEN_API
-}
